Added functionality for dividend info and improved db interactions.
Script now pulls ticker dividend info, and instead of dropping and
recreating tables each time, new dates are inserted and historical
adj_close prices are updated. This should prevent problems if other
tables use this table's date field as a foreign key.

Also changed the type of the date field to be TIMESTAMP WITH TIME ZONE
and added a time of 16:00 E[DS]T to the dates.
to timestamp without time zone and added a time of 16:00 to the dates.