Changed the way option chains are stored.
In the prior revision, it was necessary for the IB client to know
the last price of the underlying to store the option. This can
cause problems when using rt bars as we can not be sure of the
order in which we receive the data. In this revision, we specify
ahead of time which strike prices we will track, thus we are no
longer dependent on the underlying's price for data storage.